+++ bindings/python/src/gar/expose-parallel.cpp
@@ -11,7 +11,7 @@ using namespace gar;
 using context::Scalar;
 using riccati_base_t = RiccatiSolverBase<Scalar>;
 using knot_t = LqrKnotTpl<context::Scalar>;
-using lqr_t = LQRProblemTpl<context::Scalar>;
+using lqr_t = LqrProblemTpl<context::Scalar>;
 
 void exposeParallelSolver() {
 #ifdef ALIGATOR_MULTITHREADING