+++ gar/include/aligator/gar/dense-riccati.hpp
@@ -43,7 +43,7 @@ public:
   VectorXs thGrad;
   MatrixXs thHess;
 
-  explicit RiccatiSolverDense(const LQRProblemTpl<Scalar> &problem);
+  explicit RiccatiSolverDense(const LqrProblemTpl<Scalar> &problem);
 
   bool backward(const Scalar mudyn, const Scalar mueq);
 
@@ -57,7 +57,7 @@ public:
 
 protected:
   void init_factor(const KnotType &knot);
-  const LQRProblemTpl<Scalar> *problem_;
+  const LqrProblemTpl<Scalar> *problem_;
 };
 
 } // namespace aligator::gar