+++ gar/src/lqr-problem.cpp
@@ -4,12 +4,12 @@
 namespace aligator {
 namespace gar {
 template struct LqrKnotTpl<context::Scalar>;
-template struct LQRProblemTpl<context::Scalar>;
+template struct LqrProblemTpl<context::Scalar>;
 
 static_assert(
     std::uses_allocator_v<LqrKnotTpl<context::Scalar>, polymorphic_allocator>,
     "");
-static_assert(std::uses_allocator_v<LQRProblemTpl<context::Scalar>,
+static_assert(std::uses_allocator_v<LqrProblemTpl<context::Scalar>,
                                     polymorphic_allocator>,
               "");